services/flask-backend/requirements.txt
# Flask Backend - Requirements
# Python 3.13+

# Web Framework
Flask==3.1.0
flask-cors==5.0.0

# Database
pydal==20241031.1
psycopg2-binary==2.9.10

# Authentication
PyJWT==2.10.1
bcrypt==4.2.1

# Validation
email-validator==2.2.0

# WSGI Server
gunicorn==23.0.0

# Configuration
python-decouple==3.8

# Monitoring
prometheus-client==0.21.1

# Development
pytest==8.3.4
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
mypy==1.14.0
